external-services/facility_service/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic[email]
redis
httpx